Move BatchStatement to scylla-cql-core
Necessary because of From impl in scylla. After we make a release with
the split, we can re-introduce it in scylla-cql, making this one unused
(apart from From impl in scylla).
